Day 3 Lesson changes
Completed the else statement in the createList function by appending the list item to a higher list by calculating the difference between the current and previous level, creating a for loop to go up to the higher level, and then updating the value of prevLevel. Also completed the bc_switch.js script by creating a function called setupStyles that creates a link element for the page view styles, appends the link element to the document head, inserts buttons for the style switcher, and adds style rules to an embedded style sheet that is appended to the document head. These style rules are applied to div elements with the id of "styleButtons" and the input elements within them. Styles are also added to the function to turn the Page View style off and on.